Correct AST transformation for array field access from structs
The current code does not rewrite the AST as is done for function invocation made from struct fields. This simply caused the compiler to reject valid code such as x.array[0]

Fix this, cleanup the original TODO in the GC code & introduce a test. Unfortunately this test has identified another couple of issues.

The first is the already known issue of tracking spilled GC roots. This happens for calls such as x(y(), z(100)) where the result value from y() will be spilled to the stack to load the integer literal 100 into rdi in preparation for the call to z(). If z() allocates we will invoke the GC & if y() returns a pointer to the heap it won't be marked because the GC can't "see it" currently.

The second is the mis-identification of variables as free when they are not. The testcase describes the problem.