Annotate the AST with the correct symbol when rewriting free variables
Currently the old symbol is left in place which is marked as stack allocated. This causes confusion in the code generator as a stack slot is created for a field of a struct & this results in invalid values slot reads & writes for other variables.

Fix this by setting the correct symbol.