Store symtab in AST node when new scope is opened
Currently symtabs are mostly used to check for redeclares in statements which open a new "scope" of symbols. As such they mostly aren't required once that scope or block has been processed. The upcoming rework of closures will need this information to properly track which variables need captured.

As such store symtabs in the AST node which they record symbols for.