Add new AST walking code
The current code walks every type of AST node in the same way. This is leads to a confusing visit order for various types. The visit function itself is also overly complicated.

This commit introduces a pair of new walk functions to support pre & post-order walks which visits a node based on its type using a much simpler visitor function. The intention is that this will replace the existing walk code.