Remove symbol from array access AST nodes
Currently the symbol from the left node (if any) is promoted up the tree on array access. For example, the following code:

x[1] = 3

would result in the symbol 'x' being promoted to the 'opArray' AST node representing the access. This makes no sense & there is no symbol to represent the result of accessing an element of an array.

As such remove it.