Rewrite closure free variable analysis
The currently implementation accidentally captures any variable not defined in the root symtab. In practice this means it mis-identifies local variables defined inside new scopes (if, while, etc) as free as well as any field identifiers from structs. Ultimately this leads to a rather broken implementation.

Fix this by completely rewriting the logic which checks for free variables. This implementation maintains a stack of symtabs as the AST for a closures is walked in a pre-order fashion. When a variable is found the code then examines each symtab in the stack for symbol ownership & if none is found then the variable is free.

Also add a much better test closures to prevent regression.