Correct register usage when computing destination address for assignment
Currently the result of the RHS of an assignment statement is stored in rcx while the LHS address to assign it to is computed. This is not safe & rcx is used to pass the 4th parameter as per the x64 ABI.

For example:

x(1, 2, 3, 4).field = someValue()

Result in the result of 'someValue()' being overwritten by the value '4'.

Fix this by pushing the result of the RHS to the stack temporarily while the LHS address is computed.

Also add a test to prevent regression.