Remove unsafe usage to access type information during GC
Currently the compiler generates a 'value' array of typeInfo structs & uses unsafe access to generate pointers to each entry for use during GC marking. To get the starting address of this array an external function is defined (which is definitely not a function!). Apart from being a horrible hack the type info can never be used when the GC is active as it is not marked as RO.

Fix this mess by making the compiler output an array of pointers to typeInfo structs which are marked as read-only. This allows the runtime to interact with them without the use of unsafe.