Correct ASM exit sequence insertion for functions which return nothing
Currently the code generator checks if the last statement of the function is an explicit return and if so omits adding the standard ASM exit sequence. Unfortunately this function is completely wrong in that it searches for an explicit return _anywhere_ in the function AST.

Fix this by correctly examining the last statement of 'nothing' return type functions to ensure an explicitly return statement & if not insert the correct ASM instructions.

Also cleanup the codegen for functions in general.