Introduce Clara code entrypoint() to invoke main() rather than C code
Currently the C main function simply invokes the Clara main function to begin program execution. We would like to update the Clara function signature of main to be:

fn main(args: []string)

To do this we will need to parse the C main args in Clara code to allocate a new array & build the corresponding strings which need to be reachable by the GC.

The commit lays the groundwork for that by calling a new entrypoint() function which in turn invokes main(). This function will pick up the responsibility of argument parsing in a future commit.