Convert 'typeInfo' struct into an enum
Currently the typeInfo struct assumes that all types have GC roots to walk at known offsets or none. This is correct for structs, enums & simple pointer types such as strings. It is not correct for arrays which is does not currently cause a problem because arrays cannot contain pointer types. An upcoming commit is going to change that.

As such this commits prepares the way by providing support for different types of typeInfo containing different information.