Add support for find substrings to the standard library
This commit adds an Index() function which accepts a string to search and a substring to look for. It performs a simple brute force substring match on the supplied input & returns the starting index in case of a match or -1 if none is found.

A future commit will replace this will something more efficient such as Rabin-Karp or Boyer-Moore. For now this is sufficient.