Parse all environment variables on program startup
This commit adds runtime support for accessing the environment variables which were defined when the program was started. It does this by using a not POSIX-defined but an apparently widely support extension which passes a 3rd parameter to the C main() function - 'envp'. It is described in C99 as such:

"In a hosted environment, the main function receives a third argument, char *envp[], that points to a null-terminated array of pointers to char, each of which points to a string that provides information about the environment for this execution of the program"

This value is threaded into the Clara entrypoint & all env vars are parsed and stored in a new 'env' struct which is part of the 'runtime' struct which programs can access. Various functions have been added to query the environment (size, isSet, get). Due to the lack of 'null' or Option/Maybe support a horrible, buggy sentinel value is used to signify an undefined key.

References:

- https://stackoverflow.com/a/10321474
- http://port70.net/~nsz/c/c99/n1256.html#J.5.1