Update symbols to identify whether they represent a type
Currently when creating a named type the code examines the symbol table for an entry with corresponding value. However it does not check that the symbol actually represents a type (as opposed to a named variable). The result of this the the 'runtime.clara' test which should have failed typechecking but didn't

This commit fixes that by storing an additional flag in a symbol which indicates if the symbol represents an actual type.