Allow array types as expressions operands
Currently the 'unsafe' functionality can only support casting to named types because an array type is not considered a valid expression & hence cannot appear in the 3rd parameter position.

This commit changes that updates the parser to support array types as expression operands. Unfortunately function types should also be allowed but clash with anonymous functions. This has been left as a limitation for now.

This means the remaining 'casting support' provided by the C-runtime can be removed.