Support passing overloaded functions as function call arguments
Currently overloaded function definitions are permitted at global scope. During typechecking an invocation of such a function is checked against all overloads for a signature matching the supplied arguments. However attempting to _use_ an overloaded function as parameter to another function is not supported.

This commit removes that restriction and allows overloaded functions to be supplied as parameters by performing the same checking as for invocations. This allows 'println' to be passed as function unambiguously for example.