Support AST output filtering based on a regex
Currently passing '--ast' flag to the compiler results in a complete dump of the AST. This is now quite large & finding a section of interest difficult.

Fix this by update the flag to accept a regular expression argument. This allows selectively printing of top-level nodes from the AST.

For example, to match the 'main' function would be '--ast=main'. To match all functions beginning with 'copy' & 'gc' would be --ast="copy|gc". On so on.