Track in-use register types to assist GC during spill & restore
Currently argument evaluation for a function may require spilling
previously calculated value(s) to the stack. Importantly, these values
do not have stack slots like normal variables and parameters & as such
are not automatically included in the generated pointer maps. If the
spilled value is a pointer and the GC runs it will not mark it &
subsequently sweep it, likely leading to heap corruption.

This commit fixes that by tracking the types of spilled registers. When
spilling in-use registers, any pointer values are added to the current
GC map & when restoring registers those values are removed again. This
enables proper tracking of values without a "normal" stack slot.