Add basic peep hole optimiser for generated assembly
Currently the single register stack machine used for code generation
produces a number of unnecessary mov instructions which are simply due
to the mechanics of this approach.

This commit removes these instruction in two cases:

1. Simple parameter or variable loads destined for a func call must go
through rax before landing in their target register.

2. Simple parameter or variable loads destined for expression evaluation
 must go through rax before being pushed to the stack.

It does this by implementing a very simple peep hole optimiser which has
 a view on the previous & current instructions only. For instances which
  match the cases above it "squashes" the two instructions together by
  removing the temporary register load.

This change is primarily to support more readable assembly code rather
than improving performance. That said, it does result in ~10% reduction
in generated instructions. More importantly it lays the groundwork for
implementing a full linear scan register allocator in a future commit.