Switch code generator to evaluate left before right sub expressions
Currently the code generator will generate code for the right subtree of
an expression _before_ the left. This is an old holdover from when the
code generator was originally written & largely due to my inexperience
writing assembly code. Today it doesn't cause any major problems other
than occasionally being caught out by the strange evaluation order.

This commit fixes the ordering to always generate code for the left
subtree, then the right. The extra stack manipulation to achieve this is
unfortunate to do so but necessary.

Perhaps the peephole optimiser can help remove some superfluous
instructions in a later commit.