Skip typecheck if fn call arguments have already been processed
This happens in the case of a normal dot selection. For example:

```
x.println() // Rewritten to println(x)
```

At this point 'x' will have been type-checked but will be processed
again by the argument checking as part of checking the 'println' call.
This causes an issue for nested function returns where local variables
are processed twice leading to 'redeclared' errors.