Support boxing & unboxing of primitive types at generic call sites
Currently generic parameters & variables are not output in the generated
pointer maps. This means that any value not reachable from
somewhere else in the stack will be collected during the generic call if
an allocation occurs. To prevent this we need to teach the compiler that
all generic slots are in fact pointers and should be included in the
standard GC marking phase. To do this we need to support boxing.

Boxing ensures that all primitive types which find their way into a
generic context are automatically "promoted" to the heap such that the
GC marking phase can work with them. Vice versa when a generic function
returns a value, if this is bound to a primitive type, then the value
must be unboxed.

The boxing logic is implemented as a separate AST pass which runs after
typecheck. It works by analysing the types of arguments passed into &
returned from all generic call sites, boxing & unboxing values where
necessary. This is straightforward in the simple case of int →
intWrapper but quickly becomes complicated when functions which accept
or return primitives are themselves passed as arguments. In these cases
"shim" functions must be generated & invoked to generate a new
function which automatically performs boxing/unboxing for arguments &
return values which then invokes the original function ensuring all
values are a primitive or a wrapper in the correct context.

The one significant limitation of this functionality is arrays. The
current implementation simply (and sadly) panics if an attempt is
made to pass an array of primitives at a generic call site.
More generally, it is not yet clear how to support generic arrays so
leave them out for now.