Correct type reification in the default case
The existing code performs map lookup where the keys are pointers. This
provides reference equality in Go, not content equality which is what is
what was intended.

Fix that by simply iterating over the map entries and checking for a
match using the type matching functionality. This _should_ be quick as
there will likely never be more than 5 types bound for a generic method
call, struct or enum.

Drive-by change to remove unused local variable.