Add support for compile-time attributes associated with top level types
This commit introduces "attributes" as a concept into the language.
Syntactically they are a optional single list of identifiers which may
proceed any top-level type. The are not extensible & currently only a
single attribute, `ExtRet`, is defined which instructs the code
generator to omit the return address calculation required to jump over
pointer map addresses. This is required for instance when a Clara
function returns into C code.

This change paves the way for adding signal handling into the runtime.