Add basic support for timing
This commit updates the runtime to allow querying the current time via
`gettimeofday` & basic operations to convert the returned value from
micros -> millis -> seconds.

Confusingly the `timeval` argument to `gettimeofday` appears not to
match the C header files present on OSX. They declare a struct which
contains two 32-bit fields with no obvious way to allow for a 64-bit
version. That is however what happens and the declaration can
be seen in `_timeval64.h`.

This is likely a misunderstanding on my part as to how compiling libc
for 64-bit systems actually works.

For more info, see:
 - https://opensource.apple.com/source/xnu/xnu-4570.31.3/bsd/sys/_types/_user64_timeval.h.auto.html