Correct index out of bounds trampoline code
Currently the instructions following the `ioob` label are not marked as
executable meaning the binary crashes with EXC_BAD_ACCESS(code=2).
Further to this the stack in not guaranteed to be 16-byte aligned before
calling back into Clara code to print info & stop the program.

This commit fixes both of these issues.

For more info, see:
 - https://opensource.apple.com/source/xnu/xnu-201/osfmk/mach/kern_return.h.auto.html