Completely separate function invocation from function identity
Currently the AST captures both function invocation & identity in the
same AST node for global & stack identifiable functions. Other sources
of function calls such as immediate closure invocation do not have an
identity at all. This introduces two ways of determining
function information throughout the compiler & is considerably error
prone.

This commit fixes that by representing the name of a function as its
left child in the AST. This completely separates identity & invocation &
unifies the two paths through the compiler. In doing so this corrects a
long-standing hack in the parser & opens the door to introduce Pratt
parsers for easier expression handling.