Rewrite expression parsing to use Pratt parsers
This commit completely reworks expression parsing to use a Pratt parser
approach. This is significantly easier to read & crucially, extend. This
commit opens the door to easy addition of mixfix & postfix operators. It
also should allow easier rework of the parser to support passing
function types as parameters.

For more info, see:
 - http://journal.stuffwithstuff.com/2011/03/19/pratt-parsers-expression-parsing-made-easy/

Drive-by change to remove a few "fix parser" comments. In these cases
the parser was correct, my understanding was wrong. As there is no
statement separator such as a semi-colon in Clara, new line character
do not prevent subsequent lines being greedily consumed in search of the
end of an expression.

For example:

b()
(1).println()

Is actually treated as a single statement: b()(1).println()