Add support for array-based for-each loop
Currently looping is achieved through recursion or use of the 'while'
statement. The later has proved incredibly accident prone as it is too
easy to forget to increment the indexing variable at the end of the
block.

This commit addresses this by introducing a for-each statement which
operates on arrays. It has the form:

for <var> in <array> {
  // Statements
}

Where '<var>' is automatically managed to provide access to the next
element the array for every iteration of the loop.

Future commits will likely extend this functionality to
possibly provide:
 - access to index (e.g. for v, i in c)
 - integer range support (e.g. for x in 0 .. 100)
 - a predicate to support filtering (e.g. for x in c where x < 10).
 - user-defined collection support via iterators