Add support for range-based for-each loop
This commit updates the recently added for-each loop to support numeric
ranges. It has the form:

for <var> in <expr1> .. <expr2> {
  // Statements
}

Where '<var>' is automatically managed to provide access to the next
range value for every iteration of the loop. `var`is assigned value
`<expr1>` as is  incremented towards `<expr2>`, unless `<expr2>` is a
literal zero in which case this is inverted.

For example:

0 .. x → increment from 0 to x, start inclusive, end exclusive
x .. 0 → decrement x to 0, start exclusive, end inclusive
x .. y → increment from x to y, start inclusive, end exclusive

Drive-by change to correct closure capture analysis as for
statements now open a scope to examine for free variables.