Introduce built-in pointer type
Currently there is no distinction between a pointer & an integer in the
runtime. They are both represented by the same built-in type, int &
meaning is dependent on context.

An upcoming change will introduce tagging for primitive types int, byte
& bool. In such cases pointer must *never* be tagged. As such the commit
introduces a new type to ensure a clean separation between the two use
cases.