Move invokeDynamic into Clara code
Currently support for detecting & executing closures or function
pointers is implemented in assembly. However, Clara code has been able
to express this logic itself for some time.

As such ditch the assembly and move the code into the runtime. The onlyt
tricky part is making sure that invokeDynamic is not itself instrumented
when compiling the runtime. A simple check before executing the
'rewriteAnonFnAndClosures()' pass suffices to ensure this.