Add "RawValues" directive to handle tagging & tagging for external fns
Currently all external functions are assumed to require untagged
arguments, even those implemented as part of the Clara runtime which may
be in C or assembly. This leads to confusing calls to "unsafe" which
have to account for the fact that pointer is automatically adjusted
forward by 8. Also, return values are not tagged for external calls
which produce types such as int or byte which will likely lead to
crashes.

Clean all of this up by introducing a new directive named "RawValues".
Annotating an external function like this means that it requires "raw"
arguments (i.e. without Clara tags) and its return value will need
tagging. Using this information then add proper tagging for return
values as required during code gen. This removes the confusing "unsafe"
calls & allows for the runtime to take full control.

Unfortunately this directive only works for simple types which are
passed to external calls such as int, byte & arrays. Composite types
which contain int & byte values are not altered. A separate scheme will
need to be devised for these cases.