Ensure all type parameters are bound when handling return types
Currently, it is possible to invoke a generic method but fail to
specify a type parameter used in the return type. This happens when the
function call arguments do not bind "output" type parameters. In these
cases type parameters must be explicitly declared as part of the
function call.

This commit ensures that function call type checking validates that all
output type parameters are known and if not returns an error.