Report all unknown types during type instantiation
Currently, when instantiating a parameterised type if a parameter type
is unknown it is reported through callback but the partially completed
type is still returned. This leads to compiler crashes when, for
instance, a function parameter is nil.

This commit fixes this by always returning a nil type if _any_ part of
the parameterised type is problematic. This ensures the type checking
will skip over it later.