Add support for a new builtin type `bytes`
This commit introduces a builtin type representing a mutable, fixed
length sequence of bytes called `bytes`. Reads of the sequence
return an int and writes accept an int via the corresponding functions.
The intention is that this will allow the removal of the builtin `byte`
type as it seriously complicates a lot of the compiler backend

This inspiration for this comes (yet again!) for OCaml.