Output GC map root offsets as int type rather than byte
This commit switches the offsets to be ints rather than
bytes. This removes the final usages of byte type within
the language. The next commit will strip the byte type
from the compiler itself.

However, this change does increase the final binary size
by ~6Kb. We'll tackle this in a future commit by dropping
the GC maps altogether.