Remove string array operator
Currently, Clara effectively implements strings as readonly
byte arrays. This allows indexing support via square brackets
to return the byte value of a particular index in the string.
This complicates the compiler backend as a special "string read"
case is required when performing array expression evaluation.

Ditch this in favour a dedicated function to perform the
lookup. This kills yet another special case & allows complete
removal of byte & []byte from the compiler.